Conflicts:
	src/ClassicUO.Client/ClassicUO.Client.csproj
	src/ClassicUO.Client/Configuration/Profile.cs
	src/ClassicUO.Client/Configuration/Settings.cs
	src/ClassicUO.Client/Game/GameActions.cs
	src/ClassicUO.Client/Game/GameObjects/Entity.cs
	src/ClassicUO.Client/Game/GameObjects/Item.cs
	src/ClassicUO.Client/Game/Managers/AutoLootManager.cs
	src/ClassicUO.Client/Game/Managers/CommandManager.cs
	src/ClassicUO.Client/Game/Managers/DiscordManager.cs
	src/ClassicUO.Client/Game/Managers/EventSink.cs
	src/ClassicUO.Client/Game/Managers/HouseCustomizationManager.cs
	src/ClassicUO.Client/Game/Managers/MacroManager.cs
	src/ClassicUO.Client/Game/Managers/TargetManager.cs
	src/ClassicUO.Client/Game/Managers/TileMarkerManager.cs
	src/ClassicUO.Client/Game/Pathfinder.cs
	src/ClassicUO.Client/Game/Scenes/GameScene.cs
	src/ClassicUO.Client/Game/Scenes/LoginScene.cs
	src/ClassicUO.Client/Game/UI/Controls/TextBox.cs
	src/ClassicUO.Client/Game/UI/Gumps/BaseOptionsGump.cs
	src/ClassicUO.Client/Game/UI/Gumps/GridContainer.cs
	src/ClassicUO.Client/Game/UI/Gumps/GridHightlightMenu.cs
	src/ClassicUO.Client/Game/UI/Gumps/JournalGump.cs
	src/ClassicUO.Client/Game/UI/Gumps/ModernOptionsGump.cs
	src/ClassicUO.Client/Game/UI/Gumps/MultiItemMoveGump.cs
	src/ClassicUO.Client/Game/UI/Gumps/OptionsGump.cs
	src/ClassicUO.Client/Game/UI/Gumps/SpellBar/SpellQuickSearch.cs
	src/ClassicUO.Client/Game/UI/Gumps/TopBarGump.cs
	src/ClassicUO.Client/Game/UI/NearbyLootGump.cs
	src/ClassicUO.Client/GameController.cs
	src/ClassicUO.Client/LegionScripting/API.cs
	src/ClassicUO.Client/LegionScripting/Commands.cs
	src/ClassicUO.Client/LegionScripting/LegionScripting.cs
	src/ClassicUO.Client/LegionScripting/ScriptManagerGump.cs
	src/ClassicUO.Client/LegionScripting/Utility.cs
	src/ClassicUO.Client/Network/NetClient.cs
	src/ClassicUO.Client/Network/OutgoingPackets.cs
	src/ClassicUO.Client/Network/PacketHandlers.cs
	src/ClassicUO.IO/UOFileMul.cs
	src/ClassicUO.Utility/MathHelper.cs

+++ src/ClassicUO.Utility/MathHelper.cs
@@ -33,7 +33,7 @@ namespace ClassicUO.Utility
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static ulong Combine(uint val1, uint val2)
         {
-            return (val1 | (val2 << 32));
+            return (ulong)(uint)val1 | ((ulong)(uint)val2 << 32);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
